refactor(mcp): update UI labels and improve McpLlmConfig structure #371
