+++ core/src/main/kotlin/cc/unitmesh/devti/mcp/ui/model/McpLlmConfig.kt
@@ -1,9 +1,12 @@
-package cc.unitmesh.devti.mcp.ui
+package cc.unitmesh.devti.mcp.ui.model
 
 import io.modelcontextprotocol.kotlin.sdk.Tool
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 
+/**
+ * Based on https://github.com/jujumilk3/leaked-system-prompts/blob/main/anthropic-claude-api-tool-use_20250119.md
+ */
 data class McpLlmConfig(
     var temperature: Double = 0.0,
     var enabledTools: MutableList<Tool> = mutableListOf(),
@@ -30,7 +33,7 @@ String and scalar parameters should be specified as is, while lists and objects 
 
 Here are the functions available in JSONSchema format:
 <functions>
-${enabledTools.joinToString("\n") { tool -> "<function>" + Json.encodeToString(tool) } + "</function>"} }
+${enabledTools.joinToString("\n") { tool -> "<function>" + Json.Default.encodeToString(tool) } + "</function>"} }
 </functions>
 
 Answer the user's request using the relevant tool(s), if they are available. Check that all the required parameters for each tool call are provided or can reasonably be inferred from context. IF there are no relevant tools or there are missing values for required parameters, ask the user to supply these values; otherwise proceed with the tool calls. If the user provides a specific value for a parameter (for example provided in quotes), make sure to use that value EXACTLY. DO NOT make up values for or ask about optional parameters. Carefully analyze descriptive terms in the request as they may indicate required parameter values that should be included even if not explicitly quoted.
